Move WalletApi necessities to core
