+++ core-test/src/test/scala/org/bitcoins/core/api/CoinSelectorTest.scala
@@ -1,6 +1,6 @@
-package org.bitcoins.wallet.api
+package org.bitcoins.core.api
 
-import org.bitcoins.core.api.wallet.db
+import org.bitcoins.core.api.wallet.CoinSelector
 import org.bitcoins.core.api.wallet.db.{SegwitV0SpendingInfo, SpendingInfoDb}
 import org.bitcoins.core.currency._
 import org.bitcoins.core.protocol.script.ScriptPubKey
@@ -43,7 +43,7 @@ class CoinSelectorTest extends BitcoinSWalletTest {
       scriptWitness = WitnessGenerators.scriptWitness.sampleSome,
       blockHash = None
     )
-    val utxo2 = db.SegwitV0SpendingInfo(
+    val utxo2 = SegwitV0SpendingInfo(
       txid = CryptoGenerators.doubleSha256Digest.sampleSome.flip,
       state = TxoState.DoesNotExist,
       id = Some(2),
@@ -53,7 +53,7 @@ class CoinSelectorTest extends BitcoinSWalletTest {
       scriptWitness = WitnessGenerators.scriptWitness.sampleSome,
       blockHash = None
     )
-    val utxo3 = db.SegwitV0SpendingInfo(
+    val utxo3 = SegwitV0SpendingInfo(
       txid = CryptoGenerators.doubleSha256Digest.sampleSome.flip,
       state = TxoState.DoesNotExist,
       id = Some(3),
