+++ core/src/main/scala/org/bitcoins/core/api/wallet/AddUtxoResult.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.wallet.api
+package org.bitcoins.core.api.wallet
 
 import org.bitcoins.core.api.wallet.db.SpendingInfoDb
