+++ core/src/main/scala/org/bitcoins/core/api/wallet/AddressInfo.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.wallet.api
+package org.bitcoins.core.api.wallet
 
 import org.bitcoins.core.config.NetworkParameters
 import org.bitcoins.core.hd.HDPath
