+++ wallet/src/main/scala/org/bitcoins/wallet/internal/FundTransactionHandling.scala
@@ -1,6 +1,10 @@
 package org.bitcoins.wallet.internal
 
-import org.bitcoins.commons.jsonmodels.wallet.CoinSelectionAlgo
+import org.bitcoins.core.api.wallet.{
+  AddressInfo,
+  CoinSelectionAlgo,
+  CoinSelector
+}
 import org.bitcoins.core.api.wallet.db.{AccountDb, SpendingInfoDb}
 import org.bitcoins.core.consensus.Consensus
 import org.bitcoins.core.protocol.transaction._
@@ -14,7 +18,6 @@ import org.bitcoins.core.wallet.fee.FeeUnit
 import org.bitcoins.core.wallet.utxo._
 import org.bitcoins.crypto.Sign
 import org.bitcoins.keymanager.bip39.BIP39KeyManager
-import org.bitcoins.wallet.api.{AddressInfo, CoinSelector}
 import org.bitcoins.wallet.{Wallet, WalletLogger}
 
 import scala.concurrent.Future
