+++ wallet/src/main/scala/org/bitcoins/wallet/internal/UtxoHandling.scala
@@ -1,6 +1,12 @@
 package org.bitcoins.wallet.internal
 
+import org.bitcoins.core.api.wallet.AddUtxoError._
 import org.bitcoins.core.api.wallet.db._
+import org.bitcoins.core.api.wallet.{
+  AddUtxoError,
+  AddUtxoResult,
+  AddUtxoSuccess
+}
 import org.bitcoins.core.compat._
 import org.bitcoins.core.hd.HDAccount
 import org.bitcoins.core.number.UInt32
@@ -18,7 +24,6 @@ import org.bitcoins.core.protocol.transaction.{
 import org.bitcoins.core.util.{EitherUtil, FutureUtil}
 import org.bitcoins.core.wallet.utxo._
 import org.bitcoins.crypto.DoubleSha256DigestBE
-import org.bitcoins.wallet.api.{AddUtxoError, AddUtxoResult, AddUtxoSuccess}
 import org.bitcoins.wallet.{Wallet, WalletLogger}
 
 import scala.concurrent.Future
@@ -217,7 +222,6 @@ private[wallet] trait UtxoHandling extends WalletLogger {
       vout: UInt32,
       state: TxoState,
       blockHash: Option[DoubleSha256DigestBE]): Future[AddUtxoResult] = {
-    import AddUtxoError._
 
     logger.info(s"Adding UTXO to wallet: ${transaction.txId.hex}:${vout.toInt}")
